Time: 0 ms (100.00%) | Memory: 8 MB (19.13%) - LeetSync

+++ 1406-subtract-the-product-and-sum-of-digits-of-an-integer/subtract-the-product-and-sum-of-digits-of-an-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int subtractProductAndSum(int n) {
-        long int prod = 1, sum = 0;
+        int prod = 1, sum = 0;
         while (n!=0){
             int temp=n%10;
             n/=10;
